create into_dart_extra for lossy conversions
