+++ src/into_dart.rs
@@ -30,10 +30,6 @@ impl IntoDart for () {
     }
 }
 
-impl IntoDart for u32 {
-    fn into_dart(self) -> DartCObject { (self as i32).into_dart() }
-}
-
 impl IntoDart for i32 {
     fn into_dart(self) -> DartCObject {
         DartCObject {
@@ -43,10 +39,6 @@ impl IntoDart for i32 {
     }
 }
 
-impl IntoDart for u64 {
-    fn into_dart(self) -> DartCObject { (self as i64).into_dart() }
-}
-
 impl IntoDart for i64 {
     fn into_dart(self) -> DartCObject {
         DartCObject {
@@ -54,14 +46,6 @@ impl IntoDart for i64 {
             value: DartCObjectValue { as_int64: self },
         }
     }
-}
-
-impl IntoDart for i128 {
-    fn into_dart(self) -> DartCObject { self.to_string().into_dart() }
-}
-
-impl IntoDart for u128 {
-    fn into_dart(self) -> DartCObject { self.to_string().into_dart() }
 }
 
 impl IntoDart for f32 {
